Feat: service 카테고리별 검색

+++ src/main/kotlin/com/hwalaon/wezxro_server/domain/service/controller/request/AddServiceRequest.kt
@@ -8,13 +8,13 @@ import java.util.*
 data class AddServiceRequest (
 
     @NotNull
-    val providerId: Int,
+    val providerId: Long,
 
     @NotNull
-    val categoryId: Int,
+    val categoryId: Long,
 
     @NotNull
-    val apiServiceId: Int,
+    val apiServiceId: Long,
 
     @NotNull
     val name: String,
